Update server.R
Hi Fathi,

I ended up adding a couple of platforms, those with the .ht extension and those without.  I was able to test the code with pd.clariom.s.mouse data files and pd.clariom.s.mouse.ht files (that I found on the Affymetrix site).

I was unable to load the GSEA package so I couldn't test the GSEA portion of the code.  The only other issue I ran in to was with the markdown file between lines 67-101, I tried to add the new platforms but received an error saying “subscript out of bounds.”

Best,
Kristin